# ♻️ Refactor: Add findNextNonEscapedCharPosition for single-byte charset cases